+++ scripts/classes/TaxiZoneBehavior.js
@@ -1,0 +1,5 @@
+export class TaxiZoneBehavior extends foundry.data.regionBehaviors.RegionBehaviorType {
+    static LOCALIZATION_PREFIXES = ["TAXIAPP.TaxiZone"];
+    static defineSchema() { return {} }
+    static events = {};
+}